feat: sq model static require instead of dynamic
needed to play nicely with webpack
