+++ packages/graphql-transformers/src/resultLimiter/resultLimiter.js
@@ -1,7 +1,10 @@
 const { upperFirst, omit } = require('lodash');
 const paths = require('deepdash/paths');
-const { sq } = require('@venncity/sequelize-model');
+const sequelizeModel = require('@venncity/sequelize-model');
 const { BELONGS_TO_MANY } = require('./../openCrudWhereTransform/sequelizeConsts').RELATION_TYPES;
+const models = require('./../../../../test/model');
+
+const sq = sequelizeModel.sq.init(models);
 
 function shouldLimitInFetch(args) {
   return args && !containsEveryOrNone(args.where);
